requirements.txt
requests==2.24.0
requests-toolbelt==0.9.1
python-dotenv==0.20.0